zwo_fixer: create a proper header for external inclusion
