+++ zwo_fixer/zwo_fixer.cpp
@@ -1,3 +1,4 @@
+#include "internal.hpp"
 #include "zwo_fixer.hpp"
